requirements.txt
django
djangorestframework
djangorestframework-simplejwt
djoser
psycopg2
pillow
django-cors-headers
drf-yasg
django-environ
django-filter
